# Faster String#squish