--- a/src/charms/string_extensions.cr
+++ b/src/charms/string_extensions.cr
@@ -10,6 +10,45 @@ class String
   # replace newlines with a single space and convert mutiple spaces to just one
   # space.
   def squish : String
-    gsub(/[[:space:]]+/, " ").strip
+    if ascii_only?
+      squish_ascii
+    else
+      squish_unicode
+    end
+  end
+
+  # Optimized for ASCII using String#ascii_whitespace?
+  private def squish_ascii : String
+    String.build(size) do |str|
+      print_blank = false
+      each_char do |chr|
+        if chr.ascii_whitespace?
+          if print_blank
+            str << ' '
+            print_blank = false
+          end
+        else
+          print_blank = true
+          str << chr
+        end
+      end
+    end.strip
+  end
+
+  private def squish_unicode : String
+    String.build(size) do |str|
+      print_blank = false
+      each_char do |chr|
+        if chr.whitespace?
+          if print_blank
+            str << ' '
+            print_blank = false
+          end
+        else
+          print_blank = true
+          str << chr
+        end
+      end
+    end.strip
   end
 end