## Purpose
Resolves #1157 

## Description
Optimizes String#squish for both ascii and unicode text. There is some unavoidable (I think) duplication to optimize the ascii-only case. The change could be simplified to only use `Char#whitespace?` (which works for both ascii and unicode) but it's slower for ascii-only strings. However, it's still significantly faster than using a regular expression.

I replaced bench.cr with my benchmark per @jwoertink 's comment in the Github issue. It shows the difference using the ascii-only optimization (Char#ascii_whitespace? vs Char#whitespace?). Here are the results from my machine (2014 macbook):

```
     squish regex ascii-only whitespace (600 bytes)  42.17k ( 23.71µs) (± 1.57%)  5.92kB/op   8.97× slower
 squish optimized ascii-only whitespace (600 bytes) 378.22k (  2.64µs) (± 0.54%)  1.19kB/op        fastest
squish simplified ascii-only whitespace (600 bytes) 277.88k (  3.60µs) (± 1.24%)  1.19kB/op   1.36× slower

     squish regex w/ unicode whitespace (630 bytes)  63.52k ( 15.74µs) (± 0.95%)  4.48kB/op   3.39× slower
 squish optimized w/ unicode whitespace (630 bytes) 214.08k (  4.67µs) (± 1.42%)  1.41kB/op   1.01× slower
squish simplified w/ unicode whitespace (630 bytes) 215.32k (  4.64µs) (± 0.69%)  1.41kB/op        fastest
```
Lastly, there appears to be a unintended bugfix for consecutive unicode whitespace. The original regex doesn't appear to match them (libpcre issue?). The benchmark will show the difference if you run it locally.

## Checklist
* [x] - An issue already exists detailing the issue/or feature request that this PR fixes
* [x] - All specs are formatted with `crystal tool format spec src`
* [x] - Inline documentation has been added and/or updated
* [x] - Lucky builds on docker with `./script/setup`
* [x] - All builds and specs pass on docker with `./script/test`